backend/pyproject.toml
[project]
name = "emitter-research-backend"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "openpyxl>=3.1.5",
    "pandas>=2.0.0",
    "aiosqlite>=0.20.0",
    "sentence-transformers>=3.3.0",
    "faiss-cpu>=1.9.0",
    "rank-bm25>=0.2.2",
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "numpy>=1.26.0",
    "unidecode>=1.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
